refactor: 로고 컴포넌트 분리 및 컴포넌트 네이밍 변경
