+++ src/components/header/main/Header.jsx
@@ -1,12 +1,13 @@
 import { ReactComponent as ArrowSVG } from "assets/Arrow.svg";
 import Button from "components/button/Button";
 import styled from "styled-components";
+import LogoSVG from "./logo/Logo";
 
 const Wrapper = styled.header`
   width: 100%;
 `;
 
-const ButtonContainer = styled.div`
+const ButtonWrapper = styled.div`
   padding-right: 50px;
   display: flex;
   justify-content: end;
@@ -47,117 +48,20 @@ const Arrow = styled(ArrowSVG)`
 
 const Logo = styled(LogoSVG)`
   width: 100%;
-  margin: 90px auto 0;
+  margin: 70px auto 0;
 `;
 
 function Header() {
   return (
     <Wrapper>
-      <ButtonContainer>
+      <ButtonWrapper>
         <StyledButton>
           질문하러 가기
           <Arrow />
         </StyledButton>
-      </ButtonContainer>
+      </ButtonWrapper>
       <Logo />
     </Wrapper>
-  );
-}
-
-function LogoSVG({ className }) {
-  return (
-    <svg
-      className={className}
-      width="456"
-      height="180"
-      viewBox="0 0 456 180"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g clipPath="url(#clip0_1_585)">
-        <path
-          d="M202.559 130.257C257.972 162.46 313.149 170.685 325.8 148.629C338.451 126.572 303.786 82.5864 248.372 50.3831C192.959 18.18 137.782 9.95417 125.131 32.0108C112.48 54.0669 147.146 98.0535 202.559 130.257Z"
-          stroke="#959595"
-          strokeWidth="0.5"
-          strokeMiterlimit="10"
-        />
-        <path
-          d="M193.07 122.928C238.315 168.47 289.497 190.79 307.388 172.781C325.279 154.772 303.105 103.254 257.86 57.7115C212.616 12.1694 161.434 -10.1509 143.543 7.85826C125.652 25.867 147.826 77.3859 193.07 122.928Z"
-          stroke="#CACACA"
-          strokeWidth="0.5"
-          strokeMiterlimit="10"
-        />
-        <path
-          d="M27.1029 110.234C18.1618 110.234 11.4 108.321 6.81765 104.496C2.27255 100.671 0 95.0089 0 87.5089C0 79.2589 2.3098 73.2026 6.92941 69.3401C11.549 65.4776 18.3667 63.5464 27.3824 63.5464C36.9941 63.5464 44.0353 65.4589 48.5059 69.2839C52.9765 73.0714 55.2118 78.6589 55.2118 86.0464C55.2118 94.4089 52.8461 100.54 48.1147 104.44C43.3833 108.303 36.3794 110.234 27.1029 110.234ZM25.8735 90.4901H29.2824L30.923 83.2506H24.2171L25.8735 90.4901Z"
-          fill="black"
-        />
-        <path
-          d="M57.9151 109.221C57.5798 108.434 57.319 107.44 57.1327 106.24C56.9464 105.04 56.8347 103.803 56.7974 102.528C56.7602 101.215 56.7974 99.9589 56.9092 98.7589C57.0582 97.5589 57.2631 96.5651 57.5239 95.7776C58.0827 95.5526 58.6229 95.1776 59.1445 94.6526C59.7033 94.0901 60.169 93.2089 60.5415 92.0089C60.9141 90.7714 61.1004 89.0464 61.1004 86.8339C61.1004 84.4339 60.8768 82.6339 60.4298 81.4339C60.02 80.1964 59.517 79.3714 58.921 78.9589C58.3621 78.5089 57.8592 78.2839 57.4121 78.2839C57.0768 77.6089 56.8347 76.6901 56.6857 75.5276C56.5366 74.3651 56.4621 73.1276 56.4621 71.8151C56.4994 70.4651 56.6112 69.1714 56.7974 67.9339C56.9837 66.6589 57.2631 65.5714 57.6357 64.6714C60.3925 64.7839 62.87 64.8026 65.068 64.7276C67.3033 64.6526 69.52 64.5401 71.718 64.3901C73.9533 64.2401 76.3749 64.1089 78.9827 63.9964C81.6278 63.8464 84.7386 63.7714 88.3151 63.7714C93.568 63.7714 98.0014 64.2401 101.615 65.1776C105.229 66.0776 107.967 67.7464 109.83 70.1839C111.73 72.5839 112.68 76.0151 112.68 80.4776C112.68 84.3401 111.953 87.4339 110.5 89.7589C109.085 92.0839 107.036 93.7714 104.353 94.8214C101.671 95.8339 98.4484 96.3401 94.6857 96.3401C92.3013 96.3401 90.4014 96.1526 88.9857 95.7776C87.57 95.3651 86.2661 94.4651 85.0739 93.0776L83.9004 94.1464C84.869 95.3089 86.0425 96.2089 87.421 96.8464C88.7994 97.4464 90.5131 97.7839 92.5621 97.8589C92.8974 98.9089 92.8504 99.4391 93.5113 101.251C93.8457 102.167 93.5113 102.376 94.6289 104.626C95.1877 105.751 96.1564 108.038 95.7466 109.126C93.9211 108.976 87.868 108.828 84.6268 108.603C81.4229 108.378 77.7533 108.284 73.618 108.321C71.5317 108.321 69.4827 108.378 67.471 108.49C65.4592 108.603 63.6151 108.715 61.9386 108.828C60.2994 108.978 58.9582 109.109 57.9151 109.221ZM83.9004 84.4151H87.3092L89.5445 80.6464H81.6651L83.9004 84.4151Z"
-          fill="black"
-        />
-        <path
-          d="M116.615 109.221C116.317 108.434 116.074 107.44 115.888 106.24C115.702 105.04 115.59 103.803 115.553 102.528C115.553 101.215 115.59 99.9589 115.665 98.7589C115.776 97.5589 115.963 96.5651 116.224 95.7776C116.82 95.5526 117.397 95.1776 117.956 94.6526C118.515 94.0901 118.98 93.2089 119.353 92.0089C119.725 90.7714 122.147 89.0464 122.147 86.8339C122.147 84.4339 119.688 82.6339 119.241 81.4339C118.831 80.1964 118.328 79.3714 117.732 78.9589C117.174 78.5089 116.671 78.2839 116.224 78.2839C115.888 77.6089 115.646 76.6901 115.497 75.5276C115.348 74.3651 115.274 73.1276 115.274 71.8151C115.311 70.4651 115.423 69.1714 115.609 67.9339C115.795 66.6589 116.074 65.5714 116.447 64.6714C117.043 64.7089 118.012 64.7651 119.353 64.8401C120.731 64.8776 122.315 64.9339 124.103 65.0089C125.928 65.0464 127.828 65.0839 129.803 65.1214C131.777 65.1589 133.715 65.1964 135.615 65.2339C137.552 65.2339 139.284 65.2339 140.812 65.2339C144.724 65.2339 149.064 65.1776 153.832 65.0651C158.638 64.9526 163.407 64.8026 168.138 64.6151C168.324 65.3276 168.492 66.3589 168.641 67.7089C168.79 69.0589 168.902 70.5589 168.976 72.2089C169.051 73.8589 169.088 75.5276 169.088 77.2151C169.088 78.9026 169.032 80.4589 168.921 81.8839C168.809 83.2714 168.641 84.3589 168.418 85.1464C166.667 85.3714 164.729 85.5214 162.606 85.5964C160.482 85.6714 158.322 85.6714 156.124 85.5964C153.963 85.5214 151.914 85.3714 149.976 85.1464C149.529 83.3839 149.529 79.3901 149.976 77.6651H143.494V82.2776H148.691V90.0964H143.438V94.7651H149.976C149.529 93.0026 149.529 88.9901 149.976 87.2276C153.106 87.0401 156.31 86.9464 159.588 86.9464C162.867 86.9464 165.922 87.0401 168.753 87.2276C168.865 87.9026 168.958 88.9901 169.032 90.4901C169.107 91.9526 169.144 93.6214 169.144 95.4964C169.144 97.3339 169.107 99.1901 169.032 101.065C168.995 102.94 168.921 104.609 168.809 106.071C168.697 107.534 168.548 108.584 168.362 109.221C164.45 109.146 160.203 108.978 155.621 108.715C151.038 108.49 146.549 108.378 142.153 108.378C139.769 108.378 137.254 108.396 134.609 108.434C132.001 108.509 129.468 108.584 127.009 108.659C124.587 108.771 122.445 108.865 120.582 108.94C118.757 109.053 117.434 109.146 116.615 109.221Z"
-          fill="black"
-        />
-        <path
-          d="M172.97 109.221C172.672 108.434 172.43 107.44 172.244 106.24C172.058 105.04 171.946 103.803 171.909 102.528C171.909 101.215 171.946 99.9589 172.02 98.7589C172.132 97.5589 172.318 96.5651 172.579 95.7776C173.175 95.5526 173.753 95.1776 174.311 94.6526C174.87 94.0901 175.336 93.2089 175.709 92.0089C176.081 90.7714 176.267 89.0464 176.267 86.8339C176.267 84.4339 176.044 82.6339 175.597 81.4339C175.187 80.1964 174.684 79.3714 174.088 78.9589C173.529 78.5089 173.026 78.2839 172.579 78.2839C172.244 77.6089 172.002 76.6901 171.853 75.5276C171.704 74.3651 171.629 73.1276 171.629 71.8151C171.666 70.4651 171.778 69.1714 171.964 67.9339C172.151 66.6589 172.43 65.5714 172.803 64.6714C173.846 64.7464 175.057 64.8589 176.435 65.0089C177.851 65.1214 179.285 65.2339 180.738 65.3464C182.228 65.4214 183.607 65.4589 184.873 65.4589C187.593 65.4589 190.443 65.4026 193.423 65.2901C196.441 65.1401 199.943 64.9339 203.929 64.6714C203.333 67.1464 203.091 69.1714 203.203 70.7464C203.352 72.3214 204.041 73.6901 205.27 74.8526C206.5 76.0151 208.4 77.2151 210.97 78.4526L213.954 79.0151L211.697 76.9901C209.35 75.8651 207.636 74.8151 206.556 73.8401C205.512 72.8276 204.954 71.6276 204.879 70.2401C204.805 68.8151 205.047 66.9589 205.606 64.6714C207.208 64.7089 209.014 64.8589 211.026 65.1214C213.075 65.3839 215.404 65.5151 218.011 65.5151C220.321 65.5151 222.408 65.4214 224.27 65.2339C226.17 65.0089 227.884 64.8214 229.411 64.6714C229.747 65.5714 230.008 66.6589 230.194 67.9339C230.38 69.1714 230.473 70.4651 230.473 71.8151C230.473 73.1651 230.399 74.4214 230.25 75.5839C230.101 76.7464 229.877 77.6651 229.579 78.3401C229.132 78.3401 228.61 78.5651 228.014 79.0151C227.456 79.4276 226.971 80.2339 226.561 81.4339C226.152 82.6339 225.947 84.4339 225.947 86.8339C225.947 89.0464 226.133 90.7714 226.506 92.0089C226.878 93.2089 227.344 94.0901 227.903 94.6526C228.461 95.1776 229.039 95.5526 229.635 95.7776C229.896 96.5651 230.063 97.5589 230.138 98.7589C230.25 99.9589 230.287 101.215 230.25 102.528C230.25 103.803 230.157 105.04 229.97 106.24C229.784 107.44 229.542 108.434 229.244 109.221C227.754 109.034 225.872 108.828 223.6 108.603C221.364 108.415 219.036 108.321 216.614 108.321C213.485 108.321 210.449 108.434 207.506 108.659C204.562 108.884 201.377 109.071 197.95 109.221C198.471 106.746 198.676 104.665 198.564 102.978C198.453 101.29 197.95 99.8089 197.056 98.5339C196.161 97.2214 194.82 95.9839 193.032 94.8214L189.629 93.3756L192.138 96.1714C193.74 97.2589 194.913 98.4026 195.659 99.6026C196.404 100.765 196.795 102.134 196.832 103.709C196.907 105.246 196.702 107.084 196.217 109.221C194.988 109.071 193.479 108.884 191.691 108.659C189.94 108.434 187.854 108.321 185.432 108.321C183.16 108.321 180.887 108.434 178.614 108.659C176.379 108.884 174.498 109.071 172.97 109.221Z"
-          fill="black"
-        />
-        <path
-          d="M234.013 109.221C233.715 108.434 233.472 107.44 233.286 106.24C233.1 105.04 232.988 103.803 232.951 102.528C232.951 101.215 232.988 99.9589 233.063 98.7589C233.174 97.5589 233.361 96.5651 233.621 95.7776C234.217 95.5526 234.851 95.1776 235.521 94.6526C236.192 94.1276 236.825 93.2651 237.421 92.0651C238.055 90.8651 238.539 89.1776 238.874 87.0026C239.172 85.2026 239.247 83.7401 239.098 82.6151C238.949 81.4526 238.669 80.5714 238.26 79.9714C237.887 79.3339 237.477 78.9026 237.03 78.6776C236.62 78.4151 236.267 78.2839 235.968 78.2839C235.633 77.6089 235.391 76.6901 235.242 75.5276C235.093 74.3651 235.018 73.1276 235.018 71.8151C235.056 70.4651 235.167 69.1714 235.354 67.9339C235.54 66.6589 235.819 65.5714 236.192 64.6714C237.608 64.8214 239.321 64.9901 241.333 65.1776C243.382 65.3651 245.692 65.4589 248.263 65.4589C250.349 65.4589 252.547 65.4214 254.857 65.3464C257.204 65.2714 259.476 65.1776 261.674 65.0651C263.872 64.9526 265.791 64.8214 267.43 64.6714C267.207 66.0964 267.151 67.8776 267.263 70.0151C267.412 72.1526 267.989 74.2151 268.995 76.2026C269.032 74.1401 269.088 72.0776 269.163 70.0151C269.274 67.9151 269.461 66.1339 269.721 64.6714C271.286 64.7464 272.758 64.8589 274.136 65.0089C275.515 65.1214 276.949 65.2339 278.439 65.3464C279.967 65.4214 281.699 65.4589 283.636 65.4589C284.903 65.4589 286.356 65.4214 287.995 65.3464C289.634 65.2339 291.217 65.1214 292.745 65.0089C294.272 64.8589 295.502 64.7464 296.433 64.6714C296.806 65.5714 297.085 66.6589 297.271 67.9339C297.458 69.1714 297.551 70.4651 297.551 71.8151C297.588 73.1276 297.532 74.3651 297.383 75.5276C297.234 76.6901 296.992 77.6089 296.657 78.2839C296.396 78.2839 296.042 78.4151 295.595 78.6776C295.185 78.9026 294.794 79.3339 294.421 79.9714C294.086 80.5714 293.881 81.4339 293.807 82.5589C293.732 83.6464 293.881 85.0901 294.254 86.8901C294.738 89.0651 295.26 90.7714 295.818 92.0089C296.415 93.2089 297.011 94.0901 297.607 94.6526C298.24 95.1776 298.855 95.5526 299.451 95.7776C299.712 96.5651 299.879 97.5589 299.954 98.7589C300.066 99.9589 300.103 101.215 300.066 102.528C300.066 103.803 299.972 105.04 299.786 106.24C299.6 107.44 299.358 108.434 299.06 109.221C297.532 109.071 295.669 108.884 293.471 108.659C291.311 108.434 288.572 108.321 285.257 108.321C282.09 108.321 279.072 108.434 276.204 108.659C273.372 108.884 271.193 109.071 269.666 109.221C269.517 108.434 269.386 107.44 269.274 106.24C269.163 105.04 269.088 103.803 269.051 102.528C269.014 101.215 269.014 99.9589 269.051 98.7589C269.088 97.5589 269.163 96.5651 269.274 95.7776H271.23L270.671 89.4776L267.374 99.0401L256.533 94.2589L255.863 95.7776L261.954 98.5339C262.811 98.9089 263.63 99.3214 264.413 99.7714C265.195 100.184 265.866 100.746 266.424 101.459C266.983 102.171 267.356 103.146 267.542 104.384C267.766 105.584 267.728 107.196 267.43 109.221C266.722 109.146 265.735 109.053 264.468 108.94C263.202 108.865 261.786 108.771 260.221 108.659C258.694 108.584 257.111 108.509 255.471 108.434C253.869 108.359 252.361 108.321 250.945 108.321C246.81 108.321 243.308 108.434 240.439 108.659C237.608 108.884 235.466 109.071 234.013 109.221Z"
-          fill="black"
-        />
-        <path
-          d="M303.965 109.221C303.629 108.434 303.369 107.44 303.182 106.24C302.996 105.04 302.884 103.803 302.847 102.528C302.81 101.215 302.847 99.9589 302.959 98.7589C303.108 97.5589 303.313 96.5651 303.573 95.7776C304.132 95.5526 304.691 95.1776 305.25 94.6526C305.846 94.0901 306.33 93.2089 306.703 92.0089C307.075 90.7714 307.262 89.0464 307.262 86.8339C307.262 84.4339 307.038 82.6339 306.591 81.4339C306.144 80.1964 305.622 79.3714 305.026 78.9589C304.468 78.5089 303.983 78.2839 303.573 78.2839C303.238 77.6089 302.977 76.6901 302.791 75.5276C302.605 74.3651 302.512 73.1276 302.512 71.8151C302.549 70.4651 302.661 69.1714 302.847 67.9339C303.07 66.6589 303.387 65.5714 303.797 64.6714C304.915 64.7464 306.349 64.8589 308.1 65.0089C309.851 65.1589 311.695 65.2901 313.632 65.4026C315.607 65.4776 317.432 65.5151 319.109 65.5151C320.971 65.5151 322.89 65.4776 324.865 65.4026C326.839 65.2901 328.702 65.1589 330.453 65.0089C332.204 64.8589 333.638 64.7464 334.756 64.6714C335.128 65.5714 335.408 66.6589 335.594 67.9339C335.78 69.1714 335.873 70.4651 335.873 71.8151C335.873 73.1651 335.799 74.4214 335.65 75.5839C335.501 76.7464 335.259 77.6651 334.923 78.3401C334.365 78.3401 333.806 78.5651 333.247 79.0151C332.688 79.4276 332.222 80.2339 331.85 81.4339C331.477 82.6339 331.291 84.4339 331.291 86.8339C331.291 89.0464 331.477 90.7714 331.85 92.0089C332.26 93.2089 332.744 94.0901 333.303 94.6526C333.862 95.1776 334.421 95.5526 334.979 95.7776C335.277 96.5651 335.482 97.5589 335.594 98.7589C335.706 99.9589 335.743 101.215 335.706 102.528C335.669 103.803 335.557 105.04 335.371 106.24C335.184 107.44 334.923 108.434 334.588 109.221C333.47 109.109 332.036 108.978 330.285 108.828C328.534 108.715 326.672 108.603 324.697 108.49C322.723 108.378 320.804 108.321 318.941 108.321C317.227 108.321 315.421 108.378 313.521 108.49C311.658 108.603 309.87 108.715 308.156 108.828C306.479 108.978 305.082 109.109 303.965 109.221Z"
-          fill="black"
-        />
-        <path
-          d="M428.562 109.615C425.582 109.615 422.974 109.559 420.739 109.446C418.503 109.371 416.398 109.278 414.424 109.165C412.449 109.09 410.4 109.034 408.277 108.996C406.153 108.996 403.713 109.053 400.956 109.165C400.658 108.378 400.416 107.403 400.23 106.24C400.044 105.04 399.932 103.784 399.895 102.471C399.895 101.159 399.932 99.9214 400.006 98.7589C400.118 97.5589 400.304 96.5651 400.565 95.7776C401.348 95.5526 402.018 95.1401 402.577 94.5401C403.136 93.9026 403.564 92.9276 403.862 91.6151C404.16 90.3026 404.309 88.5214 404.309 86.2714C404.309 84.1339 404.123 82.5026 403.75 81.3776C403.415 80.2526 402.968 79.4651 402.409 79.0151C401.85 78.5651 401.236 78.2839 400.565 78.1714C400.193 77.0839 399.932 75.7151 399.783 74.0651C399.634 72.3776 399.634 70.6714 399.783 68.9464C399.969 67.1839 400.304 65.6839 400.789 64.4464C403.173 64.6714 405.427 64.7839 407.55 64.7839C409.711 64.7839 411.928 64.7464 414.2 64.6714C416.473 64.5589 418.95 64.4651 421.633 64.3901C424.315 64.2776 427.389 64.2214 430.853 64.2214C438.751 64.2214 444.787 65.8714 448.959 69.1714C453.169 72.4714 455.274 77.7589 455.274 85.0339C455.274 87.4714 455.05 89.8901 454.603 92.2901C454.194 94.6526 453.43 96.8839 452.312 98.9839C451.232 101.046 449.686 102.884 447.674 104.496C445.662 106.071 443.092 107.328 439.962 108.265C436.87 109.165 433.07 109.615 428.562 109.615ZM426.047 90.2089H429.456L431.692 82.9526H423.812L426.047 90.2089Z"
-          fill="black"
-        />
-        <path
-          d="M340.473 109.165C340.175 108.377 339.933 107.383 339.746 106.183C339.56 104.983 339.448 103.746 339.411 102.471C339.411 101.158 339.448 99.9022 339.523 98.7022C339.635 97.5022 339.821 96.5084 340.082 95.7209C340.678 95.4959 341.255 95.1209 341.814 94.5959C342.373 94.0334 342.839 93.1522 343.211 91.9522C343.584 90.7147 343.77 88.9897 343.77 86.7772C343.77 84.3772 343.546 82.5772 343.099 81.3772C342.69 80.1397 342.187 79.3147 341.59 78.9022C341.032 78.4522 340.529 78.2272 340.082 78.2272C339.746 77.5522 339.504 76.6334 339.355 75.4709C339.206 74.3084 339.132 73.0709 339.132 71.7584C339.169 70.4084 339.281 69.1147 339.467 67.8772C339.653 66.6022 339.933 65.5147 340.305 64.6147C341.348 64.6897 342.559 64.8022 343.938 64.9522C345.353 65.0647 346.788 65.1772 348.241 65.2897C349.731 65.3647 351.109 65.4022 352.376 65.4022C355.095 65.4022 357.945 65.3459 360.926 65.2334C363.943 65.0834 367.445 64.8772 371.432 64.6147C370.836 67.0897 370.593 69.1147 370.705 70.6897C370.854 72.2647 371.543 73.6334 372.773 74.7959C374.002 75.9584 375.902 77.1584 378.473 78.3959L381.456 78.9584L379.199 76.9334C376.852 75.8084 375.139 74.7584 374.058 73.7834C373.015 72.7709 372.456 71.5709 372.382 70.1834C372.307 68.7584 372.549 66.9022 373.108 64.6147C374.71 64.6522 376.517 64.8022 378.529 65.0647C380.578 65.3272 382.906 65.4584 385.514 65.4584C387.824 65.4584 389.91 65.3647 391.773 65.1772C393.673 64.9522 395.387 64.7647 396.914 64.6147C397.249 65.5147 397.51 66.6022 397.696 67.8772C397.883 69.1147 397.976 70.4084 397.976 71.7584C397.976 73.1084 397.901 74.3647 397.752 75.5272C397.603 76.6897 397.38 77.6084 397.082 78.2834C396.635 78.2834 396.113 78.5084 395.517 78.9584C394.958 79.3709 394.474 80.1772 394.064 81.3772C393.654 82.5772 393.449 84.3772 393.449 86.7772C393.449 88.9897 393.636 90.7147 394.008 91.9522C394.381 93.1522 394.846 94.0334 395.405 94.5959C395.964 95.1209 396.542 95.4959 397.138 95.7209C397.398 96.5084 397.566 97.5022 397.641 98.7022C397.752 99.9022 397.79 101.158 397.752 102.471C397.752 103.746 397.659 104.983 397.473 106.183C397.287 107.383 397.044 108.377 396.746 109.165C395.256 108.977 393.375 108.771 391.102 108.546C388.867 108.358 386.539 108.265 384.117 108.265C380.988 108.265 377.951 108.377 375.008 108.602C372.065 108.827 368.88 109.015 365.452 109.165C365.974 106.69 366.179 104.608 366.067 102.921C365.955 101.233 365.452 99.7522 364.558 98.4772C363.664 97.1647 362.323 95.9272 360.535 94.7647L357.131 93.3189L359.641 96.1147C361.242 97.2022 362.416 98.3459 363.161 99.5459C363.906 100.708 364.297 102.077 364.335 103.652C364.409 105.19 364.204 107.027 363.72 109.165C362.49 109.015 360.982 108.827 359.193 108.602C357.442 108.377 355.356 108.265 352.935 108.265C350.662 108.265 348.389 108.377 346.117 108.602C343.882 108.827 342 109.015 340.473 109.165Z"
-          fill="black"
-        />
-        <path
-          d="M133.312 137.61C131.473 137.61 130.067 137.106 129.095 136.099C127.791 134.862 127.139 133.081 127.139 130.755C127.139 128.382 127.791 126.601 129.095 125.411C130.067 124.403 131.473 123.899 133.312 123.899C135.152 123.899 136.557 124.403 137.53 125.411C138.828 126.601 139.477 128.382 139.477 130.755C139.477 133.081 138.828 134.862 137.53 136.099C136.557 137.106 135.152 137.61 133.312 137.61ZM135.853 134.13C136.476 133.339 136.787 132.214 136.787 130.755C136.787 129.302 136.473 128.18 135.844 127.389C135.221 126.592 134.377 126.193 133.312 126.193C132.247 126.193 131.397 126.589 130.763 127.38C130.128 128.171 129.811 129.296 129.811 130.755C129.811 132.214 130.128 133.339 130.763 134.13C131.397 134.921 132.247 135.316 133.312 135.316C134.377 135.316 135.224 134.921 135.853 134.13Z"
-          fill="black"
-        />
-        <path
-          d="M161.98 132.592H159.247V137.25H156.575V124.295H162.181C163.473 124.295 164.503 124.629 165.272 125.297C166.04 125.965 166.424 126.999 166.424 128.399C166.424 129.929 166.04 131.01 165.272 131.643C164.503 132.275 163.406 132.592 161.98 132.592ZM163.237 129.894C163.586 129.583 163.761 129.091 163.761 128.417C163.761 127.743 163.584 127.263 163.228 126.976C162.879 126.688 162.387 126.545 161.753 126.545H159.247V130.359H161.753C162.387 130.359 162.882 130.204 163.237 129.894Z"
-          fill="black"
-        />
-        <path
-          d="M192.376 131.59H186.124V134.921H193.249V137.25H183.496V124.295H192.935V126.589H186.124V129.34H192.376V131.59Z"
-          fill="black"
-        />
-        <path
-          d="M220.73 137.25H218.04L212.784 128.048V137.25H210.278V124.295H213.098L218.224 133.339V124.295H220.73V137.25Z"
-          fill="black"
-        />
-        <path
-          d="M240.683 137.25H238.177V124.295H242.089L244.429 134.481L246.752 124.295H250.62V137.25H248.114V128.487C248.114 128.235 248.117 127.884 248.122 127.433C248.128 126.976 248.131 126.624 248.131 126.378L245.695 137.25H243.084L240.666 126.378C240.666 126.624 240.669 126.976 240.674 127.433C240.68 127.884 240.683 128.235 240.683 128.487V137.25Z"
-          fill="black"
-        />
-        <path
-          d="M267.884 124.295H270.556V137.25H267.884V124.295Z"
-          fill="black"
-        />
-        <path
-          d="M298.455 137.25H295.766L290.509 128.048V137.25H288.003V124.295H290.824L295.949 133.339V124.295H298.455V137.25Z"
-          fill="black"
-        />
-        <path
-          d="M321.491 137.25H315.946V124.295H321.491C322.288 124.307 322.952 124.4 323.482 124.576C324.384 124.875 325.114 125.423 325.673 126.22C326.121 126.864 326.427 127.562 326.59 128.312C326.753 129.062 326.834 129.776 326.834 130.456C326.834 132.179 326.491 133.638 325.804 134.833C324.873 136.444 323.435 137.25 321.491 137.25ZM323.49 127.6C323.077 126.896 322.259 126.545 321.037 126.545H318.557V135H321.037C322.306 135 323.19 134.37 323.691 133.11C323.965 132.419 324.101 131.596 324.101 130.641C324.101 129.322 323.898 128.309 323.49 127.6Z"
-          fill="black"
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_1_585">
-          <rect width="456" height="180" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
   );
 }
